requirements.txt
tweepy>=4.14,<5
anthropic>=0.40,<1
discord.py>=2.3,<3
apscheduler>=3.10,<4
aiohttp>=3.9,<4
python-dotenv>=1.0,<2
aiosqlite>=0.19,<1
